+++ crates/papyrus_gateway/Cargo.toml
@@ -13,6 +13,7 @@ futures-util.workspace = true
 hyper = { workspace = true, features = ["full"] }
 jsonrpsee = { workspace = true, features = ["full"] }
 metrics.workspace = true
+papyrus_common = { path = "../papyrus_common"}
 papyrus_proc_macros = { path = "../papyrus_proc_macros"}
 papyrus_storage = { path = "../papyrus_storage" }
 starknet_writer_client = { path = "../starknet_writer_client" }
